Move Env type to types folder
