+++ src/index.ts
@@ -1,5 +1,5 @@
 import { sync } from "./services/sync";
-import { Env } from "./types";
+import { Env } from "./types/env";
 
 export default {
   async scheduled(event: ScheduledEvent, env: Env): Promise<void> {
